Fixed negative project test and moved it to a different class
